// gauvendi-platform-service-v2/src/core/enums/common.ts
export enum RestrictionConditionType {
  ClosedToStay = 'ClosedToStay',
  ClosedToArrival = 'ClosedToArrival',
  ClosedToDeparture = 'ClosedToDeparture'
}

// Restriction Enums
export enum RestrictionSourceType {
  MANUAL = 'MANUAL',
  AUTOMATED = 'AUTOMATED',
  DEFAULT = 'DEFAULT',
  PMS = 'PMS'
}
// ...
export enum HotelRestrictionCodeEnum {
  RSTR_AVAILABLE_PERIOD = 'RSTR_AVAILABLE_PERIOD',
  RSTR_CLOSE_TO_ARRIVAL = 'RSTR_CLOSE_TO_ARRIVAL',
  RSTR_CLOSE_TO_DEPARTURE = 'RSTR_CLOSE_TO_DEPARTURE',
  RSTR_CLOSE_TO_STAY_SELLABILITY = 'RSTR_CLOSE_TO_STAY_SELLABILITY',
  RSTR_CLOSE_TO_STAY = 'RSTR_CLOSE_TO_STAY',
  RSTR_LOS_MAX = 'RSTR_LOS_MAX',
  RSTR_LOS_MIN = 'RSTR_LOS_MIN',
  RSTR_MAX_ADVANCE_BOOKING = 'RSTR_MAX_ADVANCE_BOOKING',
  RSTR_MIN_ADVANCE_BOOKING = 'RSTR_MIN_ADVANCE_BOOKING',
  RSTR_MIN_LOS_THROUGH = 'RSTR_MIN_LOS_THROUGH',
  RSTR_STAY_THROUGH_DAY = 'RSTR_STAY_THROUGH_DAY'
}

export enum RestrictionEntity {
  PROPERTY = 'PROPERTY',
  ROOM_PRODUCT = 'ROOM_PRODUCT',
  SALES_PLAN = 'SALES_PLAN',
  ROOM_PRODUCT_SALES_PLAN = 'ROOM_PRODUCT_SALES_PLAN'
}

export enum HotelRestrictionSettingMode {
  NEUTRAL = 'NEUTRAL',
  PULL = 'PULL',
  PUSH = 'PUSH'
}
// ...
export enum ConnectorTypeEnum {
  APALEO = 'apaleo',
  MEWS = 'Mews',
  OHIP = 'Ohip'
}